+++ Splatter.Splatter.ps1
@@ -1,0 +1,3 @@
+Initialize-Splatter -Verb Get, Use, Find -OutputPath .\@.ps1
+Initialize-Splatter -Verb Get, Use, Find -OutputPath .\@.min.ps1 -Minify
+Initialize-Splatter -Verb Get, Use, Find -OutputPath .\@.min.gzip.ps1 -Minify -Compress